+++ src/app/inventory/dim-item-info.ts
@@ -4,7 +4,7 @@ import { tl } from 'app/i18next-t';
 import { RootState, ThunkResult } from 'app/store/types';
 import _ from 'lodash';
 import { archiveIcon, banIcon, boltIcon, heartIcon, tagIcon } from '../shell/icons';
-import { tagCleanup } from './actions';
+import { setItemNote, setItemTag, tagCleanup } from './actions';
 import { DimItem } from './item-types';
 import { itemHashTagsSelector, itemInfosSelector } from './selectors';
 import { DimStore } from './store-types';
@@ -48,16 +48,8 @@ export const tagConfig = {
   },
 };
 
-export type TagValue = keyof typeof tagConfig | 'clear' | 'lock' | 'unlock';
-
-const tagValueStrings = [...Object.keys(tagConfig), 'clear', 'lock', 'unlock'];
-
-/**
- * Helper function to check if a string is TagValue type and declare it as one.
- */
-export function isTagValue(value: string): value is TagValue {
-  return tagValueStrings.includes(value);
-}
+export type TagValue = keyof typeof tagConfig;
+export type TagCommand = TagValue | 'clear';
 
 /**
  * Priority order for which items should get moved off a character (into the vault or another character)
@@ -131,6 +123,9 @@ export function cleanInfos(stores: DimStore[]): ThunkResult {
       return;
     }
 
+    const infosWithCraftedDate = Object.values(infos).filter((i) => i.craftedDate);
+    const infosByCraftedDate = _.keyBy(infosWithCraftedDate, (i) => i.craftedDate!);
+
     // Tags/notes are stored keyed by instance ID. Start with all the keys of the
     // existing tags and notes and remove the ones that are still here, and the rest
     // should be cleaned up because they refer to deleted items.
@@ -140,6 +135,32 @@ export function cleanInfos(stores: DimStore[]): ThunkResult {
         const info = infos[item.id];
         if (info && (info.tag !== undefined || info.notes?.length)) {
           cleanupIds.delete(item.id);
+        } else {
+          // Double-check crafted items - we may have them under a different ID. If so,
+          // patch up the data by re-tagging them under the new ID.
+          if (item.craftedInfo?.craftedDate) {
+            const craftedInfo = infosByCraftedDate[item.craftedInfo.craftedDate];
+            if (craftedInfo) {
+              if (craftedInfo.tag) {
+                dispatch(
+                  setItemTag({
+                    itemId: item.id,
+                    tag: craftedInfo.tag,
+                    craftedDate: item.craftedInfo.craftedDate,
+                  })
+                );
+              }
+              if (craftedInfo.notes) {
+                dispatch(
+                  setItemNote({
+                    itemId: item.id,
+                    note: craftedInfo.notes,
+                    craftedDate: item.craftedInfo.craftedDate,
+                  })
+                );
+              }
+            }
+          }
         }
       }
     }