+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Next
 
+* DIM now loads a saved copy of your inventory even when it is offline or Bungie.net is down. The saved copy is whatever information Bungie.net last successfully provided on that device.
+  * The refresh button now has a tooltip showing how recently DIM was able to load your inventory from Bungie.net. This can help identify when DIM's view is out of date, relative to the in-game state.
+* If DIM Sync is down, the Export Backup button will save a copy of your local data instead of just failing.
+* DIM can now automatically sync an item's log state to its tag - favorite, keep, and archive tags auto lock the item, and junk or infuse tags unlock the item. This option needs to be enabled in settings, and when it's on the item tile will no longer show the lock icon for tagged items.
+* Loadouts can be filtered with buttons that appear on the list of loadouts. These buttons are populated by any hashtags you use in the loadout name or notes.
+* Crafted items will no longer lose their tags/notes or be missing from loadouts after being reshaped. This only affects items that are newly tagged or added to loadouts - crafted weapons that were already tagged or in loadouts will not be preserved when reshaping them.
+* Worked around an issue where class item mods from the fourth artifact column would be missing for some players.
+
+### Beta Only
+
+* If you add hashtags to your loadouts' names or notes, DIM will show buttons for quickly filtering down to loadouts that include that hashtag.
+
 ## 7.49.0 <span class="changelog-date">(2022-12-25)</span>
 
 * The filter help menu item is now keyboard accessible.